requirements.txt
flask
Flask-SQLAlchemy
psycopg2
flask-cors
python-dotenv